refactor: assets index파일 추가
