+++ src/assets/index.tsx
@@ -1,0 +1,21 @@
+import HamburgerSvg from '@/assets/icon/ic_hamburger.svg?react';
+import LogoSvg from '@/assets/icon/ic_gitlog_logo.svg?react';
+import WriteSvg from '@/assets/icon/ic_create.svg?react';
+import CommentSvg from '@/assets/icon/ic_comment.svg?react';
+import MeatballSvg from '@/assets/icon/ic_column_meatball.svg?react';
+import CheckSvg from '@/assets/icon/ic_toast_positive.svg?react';
+import ErrorSvg from '@/assets/icon/ic_toast_negative.svg?react';
+import PrevSvg from '@/assets/icon/ic_pagination_prev.svg';
+import NextSvg from '@/assets/icon/ic_pagination_next.svg';
+
+export {
+  HamburgerSvg,
+  LogoSvg,
+  WriteSvg,
+  CommentSvg,
+  MeatballSvg,
+  CheckSvg,
+  ErrorSvg,
+  PrevSvg,
+  NextSvg,
+};
