+++ src/components/common/Header/DetailRight.tsx
@@ -1,5 +1,4 @@
-import CommentSvg from '@/assets/icon/ic_comment.svg?react';
-import MeatballSvg from '@/assets/icon/ic_column_meatball.svg?react';
+import { CommentSvg, MeatballSvg } from '@/assets';
 import { SectionWrapper } from '@/components/common/Header/Header.styled';
 
 const DetailRight: React.FC<{ onClick?: (action: string) => void }> = ({ onClick }) => {
